requirements.txt
Flask==2.0.3
tensorflow==2.9.1
Pillow==9.2.0
numpy==1.23.0
gunicorn==20.1.0
werkzeug==2.0.3
